requirements.txt
asn1crypto==0.24.0
boto3==1.9.50
botocore==1.12.50
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
Django==2.1.3
django-api-gateway==0.0.7
django-cors-headers==2.4.0
djangorestframework==3.9.0
djangorestframework-simplejwt==3.2.3
docutils==0.14
idna==2.7
jmespath==0.9.3
mysqlclient==1.3.10
numpy==1.15.4
pandas==0.23.4
Pillow==5.3.0
pycparser==2.19
pycryptodome==3.8.2
PyJWT==1.6.4
PyMySQL==0.9.2
pyOpenSSL==18.0.0
python-dateutil==2.7.5
pytz==2018.7
reportlab==3.5.11
requests==2.20.0
s3transfer==0.1.13
six==1.11.0
sqlparse==0.3.0
urllib3==1.24.1
